chore: rename buttonStyles file for consistency purposes
