+++ src/components/Button/index.ts
@@ -1,2 +1,2 @@
 export * from "./Button";
-export * from "./utils";
+export * from "./styles";
